+++ pkg/services/authn/authn.go
@@ -153,7 +153,7 @@ type RedirectClient interface {
 // that should happen during logout and supports client specific redirect URL.
 type LogoutClient interface {
 	Client
-	Logout(ctx context.Context, user identity.Requester) (*Redirect, bool)
+	Logout(ctx context.Context, user Requester) (*Redirect, bool)
 }
 
 type PasswordClient interface {
